Move base validation file into root of clowder.yaml
