+++ clowder/clowder/yaml/validating.py
